refactor(desktop): extract AppearanceSettings into section components
Break the monolithic 432-line AppearanceSettings into focused section
components following the established TerminalSettings pattern:

- FontPreview: extracted inline preview component
- FontSettingSection: reusable editor/terminal font section (deduplicates ~120 lines)
- ThemeSection: owns theme store hooks and grid rendering
- MarkdownStyleSection: owns markdown store hooks and select dropdown
- CustomThemesSection: placeholder for upcoming feature

Use SectionList for automatic border separators, replacing fragile
conditional className logic.
