+++ apps/desktop/src/renderer/routes/_authenticated/settings/appearance/components/AppearanceSettings/components/CustomThemesSection/CustomThemesSection.tsx
@@ -1,0 +1,11 @@
+export function CustomThemesSection() {
+	return (
+		<div>
+			<h3 className="text-sm font-medium mb-2">Custom Themes</h3>
+			<p className="text-sm text-muted-foreground">
+				Custom theme import coming soon. You'll be able to import JSON theme
+				files to create your own themes.
+			</p>
+		</div>
+	);
+}
