+++ apps/desktop/src/renderer/routes/_authenticated/settings/appearance/components/AppearanceSettings/components/CustomThemesSection/index.ts
@@ -1,0 +1,1 @@
+export { CustomThemesSection } from "./CustomThemesSection";
