+++ apps/desktop/src/renderer/routes/_authenticated/settings/appearance/components/AppearanceSettings/components/FontPreview/index.ts
@@ -1,0 +1,1 @@
+export { FontPreview } from "./FontPreview";
