+++ apps/desktop/src/renderer/routes/_authenticated/settings/appearance/components/AppearanceSettings/components/FontSettingSection/index.ts
@@ -1,0 +1,1 @@
+export { FontSettingSection } from "./FontSettingSection";
