+++ apps/desktop/src/renderer/routes/_authenticated/settings/appearance/components/AppearanceSettings/components/MarkdownStyleSection/MarkdownStyleSection.tsx
@@ -1,0 +1,42 @@
+import {
+	Select,
+	SelectContent,
+	SelectItem,
+	SelectTrigger,
+	SelectValue,
+} from "@superset/ui/select";
+import {
+	type MarkdownStyle,
+	useMarkdownStyle,
+	useSetMarkdownStyle,
+} from "renderer/stores";
+
+export function MarkdownStyleSection() {
+	const markdownStyle = useMarkdownStyle();
+	const setMarkdownStyle = useSetMarkdownStyle();
+
+	return (
+		<div>
+			<h3 className="text-sm font-medium mb-2">Markdown Style</h3>
+			<p className="text-sm text-muted-foreground mb-4">
+				Rendering style for markdown files when viewing rendered content
+			</p>
+			<Select
+				value={markdownStyle}
+				onValueChange={(value) => setMarkdownStyle(value as MarkdownStyle)}
+			>
+				<SelectTrigger className="w-[200px]">
+					<SelectValue />
+				</SelectTrigger>
+				<SelectContent>
+					<SelectItem value="default">Default</SelectItem>
+					<SelectItem value="tufte">Tufte</SelectItem>
+				</SelectContent>
+			</Select>
+			<p className="text-xs text-muted-foreground mt-2">
+				Tufte style uses elegant serif typography inspired by Edward Tufte's
+				books
+			</p>
+		</div>
+	);
+}
