+++ apps/desktop/src/renderer/routes/_authenticated/settings/appearance/components/AppearanceSettings/components/MarkdownStyleSection/index.ts
@@ -1,0 +1,1 @@
+export { MarkdownStyleSection } from "./MarkdownStyleSection";
