+++ apps/desktop/src/renderer/routes/_authenticated/settings/appearance/components/AppearanceSettings/components/ThemeSection/index.ts
@@ -1,0 +1,1 @@
+export { ThemeSection } from "./ThemeSection";
